# Conflicts:
#	relearn/harness/factory/input_factory.h
#	relearn/source/io/parser/StimulusParser.h
#	relearn/source/mpi/MPINoWrapper.h
#	relearn/source/mpi/MPIWrapper.h
#	relearn/source/neurons/NetworkGraph.h
#	relearn/source/neurons/Neurons.cpp
#	relearn/source/neurons/Neurons.h
#	relearn/source/neurons/helper/AreaMonitor.cpp
#	relearn/source/neurons/input/SynapticInputCalculator.cpp
#	relearn/source/neurons/input/SynapticInputCalculator.h
#	relearn/source/neurons/input/SynapticInputCalculators.cpp
#	relearn/source/neurons/input/SynapticInputCalculators.h
#	relearn/source/neurons/models/NeuronModels.h
#	relearn/source/relearn.cpp
#	relearn/source/sim/Simulation.cpp
#	relearn/source/util/Random.h
#	relearn/test/neurons/test_neurons.cpp
#	relearn/test/neurons/test_neurons.h
#	relearn/test/parser/test_monitor_parser.cpp
#	relearn/test/synaptic_input/test_synaptic_input.cpp

+++ relearn/harness/adapter/mpi/MpiRankAdapter.h
@@ -21,13 +21,13 @@ class MPIRankAdapter {
 public:
     constexpr static int upper_bound_num_ranks = 32;
 
-    static size_t get_random_number_ranks(std::mt19937& mt) {
-        return RandomAdapter::get_random_integer<size_t>(size_t(1), upper_bound_num_ranks, mt);
+    static int get_random_number_ranks(std::mt19937& mt) {
+        return RandomAdapter::get_random_integer<int>(1, upper_bound_num_ranks, mt);
     }
 
-    static size_t get_adjusted_random_number_ranks(std::mt19937& mt) {
+    static int get_adjusted_random_number_ranks(std::mt19937& mt) {
         const auto random_rank = get_random_number_ranks(mt);
-        return round_to_next_exponent(random_rank, size_t(2));
+        return static_cast<int>(round_to_next_exponent(random_rank, 2));
     }
 
     static MPIRank get_random_mpi_rank(std::mt19937& mt) {
